+++ src/Api/Dfe.Complete.Api/Controllers/TasksDataController.cs
@@ -210,5 +210,23 @@ namespace Dfe.Complete.Api.Controllers
             await sender.Send(request, cancellationToken);
             return NoContent();
         }
+        /// <summary>
+        /// Confirm transfer project has authority to proceed task updated successfully.
+        /// </summary>
+        /// <param name="request">The update command.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanReadWriteUpdate")]
+        [HttpPatch]
+        [Route("TaskData/ConfirmTransferHasAuthorityToProceed")]
+        [SwaggerResponse(204, "Confirm transfer project has authority to proceed task updated successfully.")]
+        [SwaggerResponse(400, "Invalid request data.")]
+        [SwaggerResponse(404, "Project/User not found.")]
+        public async Task<IActionResult> UpdateConfirmTransferHasAuthorityToProceedTaskAsync(
+            [FromBody] UpdateConfirmTransferHasAuthorityToProceedTaskCommand request,
+            CancellationToken cancellationToken)
+        {
+            await sender.Send(request, cancellationToken);
+            return NoContent();
+        }
     }
 }